Allows automerging PRs in the 'unstable' state
As outline in an [issue on
octokit](octokit/octokit.net#1763), the
`unstable` state is defined as "Failing/pending commit status that is
not part of the required status checks. Merging is allowed."

Running the "Auto Merge" workflow itself induces an `unstable` state as
it is a pending check that is not part of the required status checks.
This is different than the `blocked` state which is induced by running
one of the required status checks. Either way, a merge will only be
possible if all of the required checks have passed and so it should be
safe to allow the "Auto Merge" workflow to attempt to merge the PR given
an `unstable` state.